feat: Update GitHub Actions and dependencies
- Update actions/checkout to v4
- Update actions/setup-node to v4 with Node.js version 20
- Add step to create .nojekyll file in the build process

This commit updates the GitHub Actions workflow and dependencies to improve the build and deployment process. It updates the actions/checkout and actions/setup-node actions to their latest versions, v4. It also specifies Node.js version 20 for the setup-node action. Additionally, a new step is added to create a .nojekyll file in the build process, which is necessary for GitHub Pages deployment.